+++ backends/backend-teavm/emu/com/badlogic/gdx/utils/SharedLibraryLoaderEmu.java
@@ -1,7 +1,7 @@
 package com.badlogic.gdx.utils;
 
 import com.badlogic.gdx.Gdx;
-import com.github.xpenatan.gdx.backends.teavm.assetloader.AssetLoaderListener;
+import com.github.xpenatan.gdx.backends.teavm.assetloader.AssetInstance;
 import com.github.xpenatan.gdx.backends.teavm.TeaApplication;
 import com.github.xpenatan.gdx.backends.teavm.gen.Emulate;
 import com.github.xpenatan.gdx.backends.teavm.assetloader.AssetLoader;
@@ -11,15 +11,7 @@ public class SharedLibraryLoaderEmu {
 
     public void load (String libraryName) {
         TeaApplication app = (TeaApplication)Gdx.app;
-        AssetLoader.AssetLoad assetLoader = AssetLoader.getInstance();
-        assetLoader.loadScript(false, libraryName + ".js", new AssetLoaderListener<>() {
-            @Override
-            public void onSuccess(String url, String result) {
-            }
-
-            @Override
-            public void onFailure(String url) {
-            }
-        });
+        AssetLoader assetLoader = AssetInstance.getLoaderInstance();
+        assetLoader.loadScript(libraryName + ".js", null);
     }
 }