+++ backends/backend-teavm/src/main/java/com/github/xpenatan/gdx/backends/teavm/TeaAssetPreloadListener.java
@@ -1,0 +1,7 @@
+package com.github.xpenatan.gdx.backends.teavm;
+
+import com.github.xpenatan.gdx.backends.teavm.assetloader.AssetLoader;
+
+public interface TeaAssetPreloadListener {
+    void onPreload(AssetLoader assetLoader);
+}