+++ examples/freetype/teavm/src/main/java/com/github/xpenatan/gdx/examples/teavm/BuildFreetypeTest.java
@@ -4,7 +4,6 @@ import com.github.xpenatan.gdx.backends.teavm.config.AssetFileHandle;
 import com.github.xpenatan.gdx.backends.teavm.config.TeaBuildConfiguration;
 import com.github.xpenatan.gdx.backends.teavm.config.TeaBuilder;
 import com.github.xpenatan.gdx.backends.teavm.gen.SkipClass;
-import com.github.xpenatan.gdx.examples.teavm.launcher.FreetypeTestLauncher;
 import java.io.File;
 import java.io.IOException;
 import org.teavm.tooling.TeaVMTool;