+++ extensions/asset-loader/src/main/java/com/github/xpenatan/gdx/backends/teavm/assetloader/AssetDownloader.java
@@ -1,33 +1,16 @@
 package com.github.xpenatan.gdx.backends.teavm.assetloader;
 
 /**
+ *
+ * AssetDownloader is a low level api to download any file. The full asset path should be passed.
+ * Queue counter is used when download starts and finish with error or success.
+ *
  * @author xpenatan
  */
-public class AssetDownloader {
-    private static AssetDownload instance;
-
-    private AssetDownloader() {
-    }
-
-    public static AssetDownload getInstance() {
-        return AssetDownloader.instance;
-    }
-
-    public static void setInstance(AssetDownload instance) {
-        AssetDownloader.instance = instance;
-    }
-
-    public interface AssetDownload {
-        void load(boolean async, final String url, AssetType type, AssetLoaderListener<Blob> listener);
-
-        void loadScript(boolean async, final String url, final AssetLoaderListener<String> listener);
-
-        String getHostPageBaseURL();
-
-        int getQueue();
-
-        void subtractQueue();
-
-        void addQueue();
-    }
+public interface AssetDownloader {
+    void load(boolean async, final String url, AssetType type, AssetLoaderListener<Blob> listener);
+    void loadScript(boolean async, final String url, final AssetLoaderListener<String> listener);
+    int getQueue();
+    void subtractQueue();
+    void addQueue();
 }