+++ extensions/asset-loader/src/main/java/com/github/xpenatan/gdx/backends/teavm/assetloader/AssetLoader.java
@@ -3,47 +3,45 @@ package com.github.xpenatan.gdx.backends.teavm.assetloader;
 import com.badlogic.gdx.Files.FileType;
 
 /**
+ * AssetLoader is a wrapper on AssetDownloader that adds additional functionalities
+ * like "[website]/assets" as default asset path.
+ *
  * @author xpenatan
  */
-public class AssetLoader {
+public interface AssetLoader {
 
-    private static AssetLoad instance;
+    String getAssetUrl();
 
-    private AssetLoader() {
-    }
+    String getScriptUrl();
 
-    public static AssetLoad getInstance() {
-        return AssetLoader.instance;
-    }
+    boolean isAssetInQueue(String path);
 
-    public static void setInstance(AssetLoad instance) {
-        AssetLoader.instance = instance;
-    }
+    boolean isAssetLoaded(FileType fileType, String path);
 
-    public interface AssetLoad {
-        String getAssetUrl();
+    /**
+     * Load asset and add to FileHandle system. Skip downloading if file already exists
+     */
+    void loadAsset(String path, AssetType assetType, FileType fileType);
 
-        String getScriptUrl();
+    /**
+     * Load asset and add to FileHandle system. Skip downloading if file already exists
+     */
+    void loadAsset(String path, AssetType assetType, FileType fileType, AssetLoaderListener<Blob> listener);
 
-        boolean isAssetInQueue(String path);
+    /**
+     * Load asset and add to FileHandle system. Overwrite true will update the file.
+     */
+    void loadAsset(String path, AssetType assetType, FileType fileType, AssetLoaderListener<Blob> listener, boolean overwrite);
 
-        boolean isAssetLoaded(FileType fileType, String path);
+    /**
+     * Load script and attach to html document
+     */
+    void loadScript(String path);
 
-        /**
-         * Load asset and add to FileHandle system. Skip downloading if file already exists
-         */
-        void loadAsset(boolean async, String path, AssetType assetType, FileType fileType, AssetLoaderListener<Blob> listener);
+    /**
+     * Load script and attach to html document
+     */
+    void loadScript(String path, AssetLoaderListener<String> listener);
 
-        /**
-         * Load asset and add to FileHandle system. Overwrite true will update the file.
-         */
-        void loadAsset(boolean async, String path, AssetType assetType, FileType fileType, AssetLoaderListener<Blob> listener, boolean overwrite);
-
-        /**
-         * Load script and attach to html document
-         */
-        void loadScript(boolean async, String path, AssetLoaderListener<String> listener);
-
-        int getQueue();
-    }
+    int getQueue();
 }