+++ CHANGES.md
@@ -1,5 +1,12 @@
 [-SNAPSHOT]
 
+[1.0.5]
+- Add Config Asset preloadListener
+- Update AssetLoader and AssetDownloader
+- Add AssetInstance to obtain AssetLoader or AssetDownloader
+- Fix drawing to Gdx2DPixmapEmu
+- Update Freetype emulation to fix build errors and script loading solution
+
 [1.0.4]
 - Fix music id
 - add custom gltf resource path